refactor(types): promote EnrichedMessage.SenderID to ids.UserID phantom
Why: align with project rule "ID는 Phantom Type" — int64 boundary leaks
across enricher shims defeated the type guarantee. store.UserID is already
an alias for ids.UserID, so shims return phantom directly without casts.

+++ scanner/enricher_telegram_test.go
@@ -67,7 +67,7 @@ func TestTelegramSenderShim(t *testing.T) {
 	t.Parallel()
 
 	id, name := telegramSenderShim("tg_user_xyz")
-	if id != int64(0) {
+	if id != 0 {
 		t.Errorf("id = %d; want 0", id)
 	}
 	if name != "tg_user_xyz" {
